rename `lib/server/utils` -> `lib/server/crud`
